+++ application/dash_application/views/SARS-CoV-2.py
@@ -11,7 +11,7 @@ from ..utility import sidebar_utils
 from ..utility import log_utils
 import pinery
 import logging
-from ..utility.df_manipulation import KRAKEN2_COL, BEDTOOLS_CALC_COL, BEDTOOLS_PERCENTILE_COL, SAMTOOLS_STATS_COV2_COL
+from ..utility.df_manipulation import KRAKEN2_COL, BEDTOOLS_CALC_COL, BEDTOOLS_PERCENTILE_COL
 
 logger = logging.getLogger(__name__)
 
@@ -54,7 +54,6 @@ ids = init_ids([
     'on-target-cutoff',
 
     #Graphs
-    'on-target-reads-various-bar',
     'median-coverage-scatter',
     'on-target-reads-sars-cov-2-scatter',
     'coverage-percentiles-line',
@@ -80,16 +79,6 @@ BEDTOOLS_COV_PERC_DF = util.get_bedtools_cov_perc()
 BEDTOOLS_PERCENTILE_DF = util.df_with_pinery_samples_ius(BEDTOOLS_COV_PERC_DF, pinery_samples, util.bedtools_percentile_ius_columns)
 KRAKEN2_DF = util.get_kraken2()
 
-special_columns = ['reads mapped_human', 'reads unmapped_human', 'reads mapped_covid', 'reads unmapped_covid']
-stats_merged = SAMTOOLS_STATS_COV2_HUMAN_DF.merge(SAMTOOLS_STATS_COV2_DEPLETED_DF, how="outer", on=[SAMTOOLS_STATS_COV2_COL.Barcodes, SAMTOOLS_STATS_COV2_COL.Run, SAMTOOLS_STATS_COV2_COL.Lane], suffixes=['_human', '_covid'])
-stats_merged = stats_merged[
-    special_columns + [SAMTOOLS_STATS_COV2_COL.Barcodes, SAMTOOLS_STATS_COV2_COL.Run, SAMTOOLS_STATS_COV2_COL.Lane]
-]
-stats_merged['total'] = stats_merged[special_columns].sum(axis=1)
-for c in special_columns:
-    stats_merged[c] = stats_merged[c] / stats_merged['total']
-stats_merged = util.df_with_pinery_samples_ius(stats_merged, pinery_samples, [SAMTOOLS_STATS_COV2_COL.Run, SAMTOOLS_STATS_COV2_COL.Lane, SAMTOOLS_STATS_COV2_COL.Barcodes])
-
 # Only care for Covid numbers. Be very careful about removing this, as it will make merges break
 KRAKEN2_DF = KRAKEN2_DF[KRAKEN2_DF[KRAKEN2_COL.Name] == "Severe acute respiratory syndrome coronavirus 2"]
 
@@ -97,11 +86,6 @@ BEDTOOLS_DF = util.get_bedtools_calc().merge(
     KRAKEN2_DF, how="outer",
     on=[BEDTOOLS_CALC_COL.Run, BEDTOOLS_CALC_COL.Lane, BEDTOOLS_CALC_COL.Barcodes],
     suffixes=('', "_kraken2")
-)
-BEDTOOLS_DF = BEDTOOLS_DF.merge(
-    stats_merged, how="outer",
-    on=[BEDTOOLS_CALC_COL.Run, BEDTOOLS_CALC_COL.Lane, BEDTOOLS_CALC_COL.Barcodes],
-    suffixes=('', "_samtools")
 )
 
 # Need to convert Coverage percentiles into a wide format so merge does not explode
@@ -169,27 +153,8 @@ RAW_DATA_COLUMNS = [
     BEDTOOLS_CALC_COL.MeanCoverage,
     BEDTOOLS_CALC_COL.CoverageUniformity,
     KRAKEN2_COL.PercentAtClade,
-    # Custom samtools columns due to merge
-    'reads mapped_human',
-    'reads unmapped_human',
-    'reads mapped_covid',
-    'reads unmapped_covid',
     # Add columns produced by wide coverage percentage dataframe
 ] + [x for x in BEDTOOLS_COV_PERC_WIDE_DF if x.startswith('Coverage Above')]
-
-def generate_on_target_reads_bar(current_data, graph_params):
-    return generate_bar(
-        current_data,
-        ['reads mapped_human', 'reads mapped_covid'],
-        lambda d: d[PINERY_COL.SampleName] + d[PINERY_COL.LaneNumber].astype(str) + d[PINERY_COL.SequencerRunName],
-        lambda d, col: d[col] * 100,
-        "On-Target (%)",
-        "%",
-        fill_color={
-            "reads mapped_human": "black",
-            "reads mapped_covid": "red",
-        },
-    )
 
 
 def generate_median_coverage_scatter(current_data, graph_params):
@@ -394,9 +359,6 @@ def layout(query_string):
                             core.Graph(id=ids['coverage-uniformity-scatter'],
                                        figure=generate_coverage_uniformity_scatter(df, initial)
                                        ),
-                            core.Graph(id=ids['on-target-reads-various-bar'],
-                                       figure=generate_on_target_reads_bar(df, initial)
-                                       ),
                             core.Graph(id=ids['coverage-percentiles-line'],
                                        figure=generate_coverage_percentiles_line(percentile_df, initial)
                                        ),
@@ -429,7 +391,6 @@ def init_callbacks(dash_app):
         [
             Output(ids["approve-run-button"], "href"),
             Output(ids["approve-run-button"], "style"),
-            Output(ids['on-target-reads-various-bar'], 'figure'),
             Output(ids['median-coverage-scatter'], 'figure'),
             Output(ids['on-target-reads-sars-cov-2-scatter'], 'figure'),
             Output(ids['coverage-percentiles-line'], 'figure'),
@@ -520,7 +481,6 @@ def init_callbacks(dash_app):
         return [
             approve_run_href,
             approve_run_style,
-            generate_on_target_reads_bar(df, graph_params),
             generate_median_coverage_scatter(df, graph_params),
             generate_on_target_reads_scatter(df, graph_params),
             generate_coverage_percentiles_line(percentile_df, graph_params),